deepen: consolidate request contract and split guidance triple-seam
- Candidate 06 (Speculative): fold edit-normalize (13 lines, shallow)
  and schema (38 lines, duplicated literals) into contract.ts — one
  source for field-set validity, drift-proof. Delete shallow wrappers;
  tools now import normalizeRequest + schemas from contract.

- Candidate 05 (Worth exploring): split guidance.ts (485 lines, 3
  lifecycles) into guidance/parse (pure fence state machine),
  guidance/resolve (IO-light, first-candidate-wins), and
  guidance/materialize (IO-heavy, idempotent seeding). Barrel at
  src/guidance.ts preserves import path.

Deletion tests: edit-normalize moves, schema interface≈implementation,
guidance triple-seam concentrates fence invariant when split.
